Remove compiler warning.

+++ src/tool-tassoize.cpp
@@ -445,7 +445,7 @@ void Tool_tassoize::deleteBreaks(HumdrumFile& infile) {
 void Tool_tassoize::addBibliographicRecords(HumdrumFile& infile) {
 	std::vector<HumdrumLine*> refinfo = infile.getReferenceRecords();
 	std::map<string, HumdrumLine*> refs;
-	for (int i=0; i<refinfo.size(); i++) {
+	for (int i=0; i<(int)refinfo.size(); i++) {
 		string key = refinfo[i]->getReferenceKey();
 		refs[key] = refinfo[i];
 	}
